Move complete_linear_regression_dataset to a util.py specific to learners
This function so far is only used by ERR and was located in the dataset_reader package.
